requirements.txt
cmsplugin-gallery==0.5.5
cmsplugin-image-gallery==0.6.2
cmsplugin-nivoslider==0.5.0
dj-database-url==0.3.0
Django==1.6.8
django-admin-jqueryui==1.8.22
django-allauth==0.18.0
django-classy-tags==0.5.2
django-cms==3.0.7
django-filer==0.9.8
django-form-designer==0.8.0
django-inline-ordering==1.0.2
django-mptt==0.6.1
django-picklefield==0.3.1
django-polymorphic==0.6.1
django-reversion==1.8.5
django-sekizai==0.7
django-template-debug==0.3.5
django-template-repl==0.3.0
django-treebeard==2.0
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.6.0
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
easy-thumbnails==2.2
html5lib==0.999
oauthlib==0.7.2
Pillow==2.6.1
-e git+http://git@github.com/divio/django-polls.git@4941bdd935a7fc01518ba0dbc3eef8e896254a0e#egg=polls-master
psycopg2==2.5.4
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.8.0
sorl-thumbnail==12.2
South==1.0.2
Unidecode==0.4.17
